requirements.txt
amqp==5.3.1
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
babel==2.17.0
backcall==0.2.0
beautifulsoup4==4.13.4
billiard==4.2.1
bleach==6.3.0
branca==0.8.1
Brotli==1.1.0
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==45.0.3
cssselect2==0.8.0
daphne==4.2.0
decorator==5.2.1
defusedxml==0.7.1
Django==5.0.6
django-admin-interface==0.28.8
django-ckeditor==6.7.2
django-colorfield==0.11.0
django-cors-headers==4.4.0
django-extensions==3.2.3
django-filter==25.1
django-js-asset==3.1.2
django-tinymce==4.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
docopt==0.6.2
docxcompose==1.4.0
docxtpl==0.20.1
drf-yasg==1.21.10
et_xmlfile==2.0.0
executing==2.2.1
Faker==38.0.0
fastjsonschema==2.21.2
folium==0.19.5
fonttools==4.58.0
html5lib==1.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
kombu==5.5.3
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib-inline==0.2.1
mistune==3.1.4
mysql==0.0.3
mysql-connector-python==9.5.0
mysqlclient==2.2.7
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
numpy==2.2.5
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.5
pickleshare==0.7.5
pillow==11.2.1
pipreqs==0.5.0
platformdirs==4.5.0
prompt_toolkit==3.0.51
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydyf==0.11.0
Pygments==2.19.2
PyJWT==2.9.0
PyMySQL==1.1.1
pyOpenSSL==25.1.0
PyPDF2==3.0.1
pyphen==0.17.2
pypng==0.20220715.0
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.2
pyzmq==27.1.0
qrcode==7.4.2
referencing==0.37.0
reportlab==4.2.2
requests==2.32.3
rpds-py==0.28.0
service-identity==24.2.0
six==1.16.0
soupsieve==2.7
sqlparse==0.5.0
stack-data==0.6.3
tabulate==0.9.0
text-unidecode==1.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tomli==2.3.0
tornado==6.5.2
traitlets==5.14.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
xyzservices==2025.4.0
yarg==0.1.9
zope.interface==7.2
zopfli==0.2.3.post1
